Split desktop query boundaries
