+++ src/app/query/desktop-extension-query.ts
@@ -1,0 +1,107 @@
+import type {
+  PiConfiguredPackage,
+  PiConfiguredSkill,
+  PiPackageCatalogPage,
+  PiPackageMutationResult,
+  PiSkillCatalogPage,
+  PiSkillMutationResult,
+  SkillCreatorSessionState,
+} from '../desktop/types'
+
+export async function searchPiPackagesQuery(
+  request: {
+    query?: string | null | undefined
+    cursor?: number | null | undefined
+    pageSize?: number | null | undefined
+  } = {},
+): Promise<PiPackageCatalogPage> {
+  return (
+    (await window.piDesktop?.searchPiPackages?.(request)) ?? {
+      query: request.query?.trim() ?? '',
+      sort: 'monthlyDownloads-desc',
+      total: 0,
+      nextCursor: null,
+      items: [],
+    }
+  )
+}
+
+export async function searchPiSkillsQuery(
+  request: { query?: string | null | undefined; limit?: number | null | undefined } = {},
+): Promise<PiSkillCatalogPage> {
+  return (
+    (await window.piDesktop?.searchPiSkills?.(request)) ?? {
+      query: request.query?.trim() ?? '',
+      total: 0,
+      items: [],
+    }
+  )
+}
+
+export async function getConfiguredPiPackagesQuery(
+  request: { projectPath?: string | null | undefined; chat?: boolean | undefined } = {},
+): Promise<PiConfiguredPackage[]> {
+  return (await window.piDesktop?.getConfiguredPiPackages?.(request)) ?? []
+}
+
+export async function installPiPackageQuery(request: {
+  source: string
+  kind?: 'npm' | 'git' | undefined
+  local?: boolean | undefined
+  projectPath?: string | null
+  chat?: boolean | undefined
+}): Promise<PiPackageMutationResult | null> {
+  return (await window.piDesktop?.installPiPackage?.(request)) ?? null
+}
+
+export async function removePiPackageQuery(request: {
+  source: string
+  local?: boolean | undefined
+  projectPath?: string | null
+  chat?: boolean | undefined
+}): Promise<PiPackageMutationResult | null> {
+  return (await window.piDesktop?.removePiPackage?.(request)) ?? null
+}
+
+export async function getConfiguredPiSkillsQuery(
+  request: { projectPath?: string | null | undefined; chat?: boolean | undefined } = {},
+): Promise<PiConfiguredSkill[]> {
+  return (await window.piDesktop?.getConfiguredPiSkills?.(request)) ?? []
+}
+
+export async function installPiSkillQuery(request: {
+  source: string
+  local?: boolean | undefined
+  projectPath?: string | null
+  chat?: boolean | undefined
+}): Promise<PiSkillMutationResult | null> {
+  return (await window.piDesktop?.installPiSkill?.(request)) ?? null
+}
+
+export async function removePiSkillQuery(request: {
+  installedPath: string
+  projectPath?: string | null
+  chat?: boolean | undefined
+}): Promise<PiSkillMutationResult | null> {
+  return (await window.piDesktop?.removePiSkill?.(request)) ?? null
+}
+
+export async function startSkillCreatorSessionQuery(request: {
+  prompt: string
+  local?: boolean | undefined
+  projectPath?: string | null
+  chat?: boolean | undefined
+}): Promise<SkillCreatorSessionState | null> {
+  return (await window.piDesktop?.startSkillCreatorSession?.(request)) ?? null
+}
+
+export async function continueSkillCreatorSessionQuery(request: {
+  sessionId: string
+  prompt: string
+}): Promise<SkillCreatorSessionState | null> {
+  return (await window.piDesktop?.continueSkillCreatorSession?.(request)) ?? null
+}
+
+export async function closeSkillCreatorSessionQuery(sessionId: string): Promise<void> {
+  await window.piDesktop?.closeSkillCreatorSession?.(sessionId)
+}
